requirements.txt
ipdb
scipy
transformers
sentencepiece
datasets
tqdm
protobuf
promptsource
evaluate
scikit-learn
wandb
gpustat
black
open-clip-torch